# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "North Bound Interface (NBI) REST API exposed by voltha-go-controller (VGC).\nIt provides configuration and operational endpoints for subscribers, bandwidth\nprofiles, IGMP proxy, multicast, flows, devices, ports, meters, groups, DHCP\nallocations, MAC learning and the OLT flow service.",
        "title": "VOLTHA Go Controller NBI REST API",
        "contact": {
            "name": "Open Networking Foundation",
            "url": "https://opennetworking.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/vgc/v1",
    "paths": {
...
        "/programmed-subscribers": {
            "get": {
                "description": "Retrieve programmed subscribers, optionally filtered by device and port.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscribers"
                ],
                "summary": "Get programmed subscribers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.SubscribersList"
                        }
                    },
                    "404": {
                        "description": "Subscribers not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/programmed-subscribers/{device}/{port}": {
            "get": {
                "description": "Retrieve programmed subscribers, optionally filtered by device and port.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscribers"
                ],
                "summary": "Get programmed subscribers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device identifier",
                        "name": "device",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Port identifier",
                        "name": "port",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.SubscribersList"
                        }
                    },
                    "404": {
                        "description": "Subscribers not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/services/{device}/{port}": {
            "get": {
                "description": "Retrieve programmed subscribers, optionally filtered by device and port.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscribers"
                ],
                "summary": "Get programmed subscribers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device identifier",
                        "name": "device",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Port identifier",
                        "name": "port",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.SubscribersList"
                        }
                    },
                    "404": {
                        "description": "Subscribers not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Activate the service(s) on the given device and port.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Activate a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device identifier",
                        "name": "device",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Port identifier",
                        "name": "port",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service activated"
                    }
                }
            },
            "delete": {
                "description": "Deactivate the service(s) on the given device and port.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Deactivate a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device identifier",
                        "name": "device",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Port identifier",
                        "name": "port",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deactivated"
                    }
                }
            }
        },
        "/services/{portName}": {
            "post": {
                "description": "Activate the service on the given access port, optionally scoped by S-Tag, C-Tag and technology profile id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Activate a service by port name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access port name",
                        "name": "portName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service activated"
                    }
                }
            },
            "delete": {
                "description": "Deactivate the service on the given access port, optionally scoped by S-Tag, C-Tag and technology profile id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Deactivate a service by port name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access port name",
                        "name": "portName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deactivated"
                    }
                }
            }
        },
        "/services/{portName}/{sTag}/{cTag}/{tpId}": {
            "post": {
                "description": "Activate the service on the given access port, optionally scoped by S-Tag, C-Tag and technology profile id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Activate a service by port name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access port name",
                        "name": "portName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service VLAN (S-Tag)",
                        "name": "sTag",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Customer VLAN (C-Tag)",
                        "name": "cTag",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Technology profile id",
                        "name": "tpId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service activated"
                    }
                }
            },
            "delete": {
                "description": "Deactivate the service on the given access port, optionally scoped by S-Tag, C-Tag and technology profile id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Deactivate a service by port name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access port name",
                        "name": "portName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service VLAN (S-Tag)",
                        "name": "sTag",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Customer VLAN (C-Tag)",
                        "name": "cTag",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Technology profile id",
                        "name": "tpId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deactivated"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "onosnbi.SubscriberInfo": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "tagInfo": {
                    "$ref": "#/definitions/onosnbi.UniTagInformation"
                }
            }
        },
        "onosnbi.SubscribersList": {
            "type": "object",
            "properties": {
                "subscribers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.SubscriberInfo"
                    }
                }
            }
        },
...
        "onosnbi.UniTagInformation": {
            "type": "object",
            "properties": {
                "configuredMacAddress": {
                    "type": "string"
                },
                "downstreamBandwidthProfile": {
                    "type": "string"
                },
                "downstreamOltBandwidthProfile": {
                    "type": "string"
                },
                "dsPonCTagPriority": {
                    "type": "integer"
                },
                "dsPonSTagPriority": {
                    "type": "integer"
                },
                "enableMacLearning": {
                    "type": "boolean"
                },
                "isDhcpRequired": {
                    "type": "boolean"
                },
                "isIgmpRequired": {
                    "type": "boolean"
                },
                "isPppoeRequired": {
                    "type": "boolean"
                },
                "ponCTag": {
                    "type": "integer"
                },
                "ponSTag": {
                    "type": "integer"
                },
                "serviceName": {
                    "type": "string"
                },
                "technologyProfileId": {
                    "type": "integer"
                },
                "uniTagMatch": {
                    "type": "integer"
                },
                "upstreamBandwidthProfile": {
                    "type": "string"
                },
                "upstreamOltBandwidthProfile": {
                    "type": "string"
                },
                "usPonCTagPriority": {
                    "type": "integer"
                },
                "usPonSTagPriority": {
                    "type": "integer"
                }
            }
        }
    }
}